thomas/AF-532-refactor-stargateAM

+++ test/fuzz/asset-modules/StargateAssetModule/Stake.fuzz.t.sol
@@ -23,7 +23,7 @@ contract Stake_StargateAssetModule_Fuzz_Test is StargateAssetModule_Fuzz_Test {
                               TESTS
     /////////////////////////////////////////////////////////////// */
 
-    function testFuzz_success_stake(uint256 amount, uint256 poolId) public {
+    function testFuzz_success_stake(uint256 amount, uint96 poolId) public {
         // Given : Random Stargate pool address.
         address poolLpToken = address(new ERC20Mock("stakingToken", "STK", 0));
